# Optimize generic collection Equal