--- a/Src/FluentAssertions/Execution/GivenSelectorExtensions.cs
+++ b/Src/FluentAssertions/Execution/GivenSelectorExtensions.cs
@@ -19,10 +19,10 @@ namespace FluentAssertions.Execution
             this GivenSelector<ICollection<T>> givenSelector, int length)
         {
             return givenSelector
-                .ForCondition(items => (items.Any() || (length == 0)))
+                .ForCondition(items => ((items.Count > 0) || (length == 0)))
                 .FailWith("but found empty collection.")
                 .Then
-                .ForCondition(items => (!items.Any() || (length > 0)))
+                .ForCondition(items => ((items.Count == 0) || (length > 0)))
                 .FailWith("but found {0}.", items => items);
         }