--- a/Tests/Benchmarks/Program.cs
+++ b/Tests/Benchmarks/Program.cs
@@ -6,7 +6,7 @@ namespace Benchmarks
     {
         public static void Main()
         {
-            var summary = BenchmarkRunner.Run<BeEquivalentToBenchmarks>();
+            var summary = BenchmarkRunner.Run<CollectionEqualBenchmarks>();
         }
     }
 }